chore: refactoring of folder names
