+++ .github/workflows/test_frameworks.yml
@@ -115,7 +115,7 @@ jobs:
         if: needs.needs-run.outputs.outcome == 'success'
         # Disable all test_simple tests because they check for
         # log output and it contains phony error messages.
-        run: PYTHONPATH=../ddtrace/tests/internal/bytecode_injection/injection/ ddtrace-run pytest test --continue-on-collection-errors -v -k 'not test_simple'
+        run: PYTHONPATH=../ddtrace/tests/internal/bytecode_injection/framework_injection/ ddtrace-run pytest test --continue-on-collection-errors -v -k 'not test_simple'
       - name: Bytecode injection result
         if: needs.needs-run.outputs.outcome == 'success'
         run: cat bytecode-injection.txt
@@ -246,7 +246,7 @@ jobs:
       - name: Run Bytecode injection tests
         if: needs.needs-run.outputs.outcome == 'success'
         # django.tests.requests module interferes with requests library patching in the tracer -> disable requests patch
-        run: PYTHONPATH=../ddtrace/tests/internal/bytecode_injection/injection/:. DD_PATCH_MODULES=unittest:no DD_TRACE_REQUESTS_ENABLED=0 ddtrace-run tests/runtests.py --parallel 1
+        run: PYTHONPATH=../ddtrace/tests/internal/bytecode_injection/framework_injection/:. DD_PATCH_MODULES=unittest:no DD_TRACE_REQUESTS_ENABLED=0 ddtrace-run tests/runtests.py --parallel 1
 
       - name: Bytecode injection results
         if: needs.needs-run.outputs.outcome == 'success'
@@ -395,7 +395,7 @@ jobs:
       - name: Bytecode injection Test
         if: needs.needs-run.outputs.outcome == 'success'
         # https://github.com/tiangolo/fastapi/pull/10876
-        run: PYTHONPATH=../ddtrace/tests/internal/bytecode_injection/injection/ ddtrace-run pytest -p no:warnings tests -k 'not test_warn_duplicate_operation_id'
+        run: PYTHONPATH=../ddtrace/tests/internal/bytecode_injection/framework_injection/ ddtrace-run pytest -p no:warnings tests -k 'not test_warn_duplicate_operation_id'
       - name: Bytecode injection results
         if: needs.needs-run.outputs.outcome == 'success'
         run: cat bytecode-injection.txt
