+++ tests/internal/bytecode_injection/framework_injection/output.py
