+++ tests/internal/bytecode_injection/framework_injection/sitecustomize.py
