+++ tests/internal/bytecode_injection/framework_injection/utils.py
