+++ tests/internal/bytecode_injection/test_injection.py
@@ -4,11 +4,11 @@ import sys
 import mock
 import pytest
 
-from ddtrace.internal.injection import InvalidLine
-from ddtrace.internal.injection import eject_hook
-from ddtrace.internal.injection import eject_hooks
-from ddtrace.internal.injection import inject_hook
-from ddtrace.internal.injection import inject_hooks
+from ddtrace.internal.bytecode_injection import InvalidLine
+from ddtrace.internal.bytecode_injection import eject_hook
+from ddtrace.internal.bytecode_injection import eject_hooks
+from ddtrace.internal.bytecode_injection import inject_hook
+from ddtrace.internal.bytecode_injection import inject_hooks
 from ddtrace.internal.utils.inspection import linenos
