+++ tests/suitespec.yml
@@ -58,7 +58,7 @@ components:
     - ddtrace/internal/logger.py
     - ddtrace/internal/hostname.py
     - ddtrace/internal/http.py
-    - ddtrace/internal/injection.py
+    - ddtrace/internal/bytecode_injection/*
     - ddtrace/internal/logger.py
     - ddtrace/internal/metrics.py
     - ddtrace/internal/module.py
